Create helloWorld.py
Navigate to the desired directory in the terminal and execute the following command to run the "helloWorld.py" Python script:

cd /path/to/your/directory && python helloWorld.py

Replace "/path/to/your/directory" with the actual path to the directory where your "helloWorld.py" file is located. This command ensures that you change to the specified directory before running the Python script.